backend/requirements.txt
Django>=6.0,<7
djangorestframework>=3.15,<4
django-cors-headers>=4.6,<5
djangorestframework-simplejwt>=5.4,<6
Pillow>=11.0,<12
celery[redis]>=5.4,<6
redis>=5.0,<6
httpx>=0.28,<1
cryptography>=44.0,<46
python-dotenv>=1.0,<2
boto3>=1.35,<2
django-storages>=1.14,<2
google-genai>=1.73,<2
google-auth>=2.0,<3
numpy>=2.0,<3
